feat: Support crowd dummy for training mode
