+++ StarResonanceDpsAnalysis.WPF/ViewModels/PersonalDpsViewModel.cs
@@ -23,7 +23,12 @@ public enum DummyTargetType
     /// <summary>
     /// T木桩
     /// </summary>
-    TDummy
+    TDummy,
+
+    /// <summary>
+    /// AOE木桩（适合测试范围伤害）
+    /// </summary>
+    AoeDummy,
 }
 
 public partial class PersonalDpsViewModel : BaseDispatcherSupportViewModel
@@ -108,59 +113,6 @@ public partial class PersonalDpsViewModel : BaseDispatcherSupportViewModel
         _configManager.CurrentConfig.DefaultDummyTarget = value;
         // 异步保存配置
         _ = _configManager.SaveAsync();
-    }
-
-    [RelayCommand]
-    private void Loaded()
-    {
-        if (_isLoaded)
-        {
-            return;
-        }
-
-        _isLoaded = true;
-
-        // 订阅配置更新事件以响应主题颜色变化
-        _configManager.ConfigurationUpdated += OnConfigurationUpdated;
-        _engine.ProcessedDataReady += EngineOnProcessedDataReady;
-
-        // ⭐ 从配置加载上次选择的木桩类型
-        var savedDummyTarget = _configManager.CurrentConfig.DefaultDummyTarget;
-        if (Enum.IsDefined(typeof(DummyTargetType), savedDummyTarget))
-        {
-            SelectedDummyTarget = savedDummyTarget;
-            _logger?.LogInformation("从配置加载木桩类型: {Type}", SelectedDummyTarget);
-        }
-
-        _engine.Configure(new DataSourceEngineParam()
-        {
-            Mode = DataSourceMode.DummyTraining,
-            DummyTarget = SelectedDummyTarget,
-            PlayerUid = _configManager.CurrentConfig.Uid > 0 ? _configManager.CurrentConfig.Uid : -1,
-        });
-
-        StartTimer();
-        _engine.CurrentSource.Refresh();
-    }
-
-    [RelayCommand]
-    private void UnLoaded()
-    {
-        if (!_isLoaded)
-        {
-            return;
-        }
-
-        _isLoaded = false;
-
-        _configManager.ConfigurationUpdated -= OnConfigurationUpdated;
-        _engine.ProcessedDataReady -= EngineOnProcessedDataReady;
-        _engine.Configure(new DataSourceEngineParam()
-        {
-            Mode = AppConfig.DpsUpdateMode.ToDataSourceMode(),
-            ActiveUpdateInterval = AppConfig.DpsUpdateInterval,
-        });
-        StopTimer();
     }
 
     private void EngineOnProcessedDataReady(object? sender,
